refactor(design-system): move shared components to common subfolder
