+++ apps/web/components/layout/footer.tsx
@@ -1,6 +1,6 @@
 import { MapsIcon } from '@hugeicons/core-free-icons';
 import { HugeiconsIcon } from '@hugeicons/react';
-import { SectionContainer } from '@repo/design-system/components/ui/section-container';
+import { SectionContainer } from '@repo/design-system/components/common/section-container';
 
 export function Footer() {
   return (
@@ -13,7 +13,7 @@ export function Footer() {
         <div className="flex items-center gap-2">
           <HugeiconsIcon className="text-primary-foreground size-8" icon={MapsIcon} />
           <span className="font-heading text-primary-foreground text-2xl font-bold tracking-[-0.5px]">
-            rmap
+            RMap
           </span>
         </div>
         <p className="flex items-center justify-center pt-1 text-sm text-slate-400">
