+++ packages/design-system/components/common/animated-icon-swap.tsx
